requirements.txt
channels==2.1.1
djangorestframework==3.8.2
